Fix bathroom parsing, add decimal support for room counts, optimize memory usage
- Fix bathroom parsing: restrict data-testid selector to avoid false matches (was always 61)
- Add decimal support: bedrooms, bathrooms, and beds now support fractional values (e.g., 1.5)
- Memory optimizations: release HTML buffers, clear large slices, kill browser processes, use /tmp/air-data for browser data, add explicit GC calls

+++ db/db.go
@@ -130,9 +130,9 @@ func (db *DB) initSchema() error {
 			status VARCHAR(20) NOT NULL DEFAULT 'pending',
 			is_superhost BOOLEAN,
 			is_guest_favorite BOOLEAN,
-			bedrooms INTEGER,
-			bathrooms INTEGER,
-			beds INTEGER,
+			bedrooms DOUBLE PRECISION,
+			bathrooms DOUBLE PRECISION,
+			beds DOUBLE PRECISION,
 			description TEXT,
 			house_rules TEXT,
 			newest_review_date TIMESTAMP,
@@ -194,4 +194,3 @@ func (db *DB) initSchema() error {
 func (db *DB) GetConn() *sql.DB {
 	return db.conn
 }
-
